+++ src/StardewModdingAPI/ISemanticVersion.cs
@@ -4,6 +4,9 @@ namespace StardewModdingAPI
 {
     /// <summary>A semantic version with an optional release tag.</summary>
     public interface ISemanticVersion : IComparable<ISemanticVersion>
+#if !SMAPI_1_x
+        , IEquatable<ISemanticVersion>
+#endif
     {
         /*********
         ** Accessors