+++ release-notes.md
@@ -11,6 +11,7 @@ For mod developers:
 * Added `InputEvents` which combine keyboard + mouse + controller input for simpler input handling; add metadata like the cursor position and grab tile for simpler click handling; and add an option to suppress input from the game.
 * Added support for optional dependencies.
 * Added support for string versions (like `"1.0-alpha"`) in `manifest.json`.
+* Added `IEquatable<ISemanticVersion>` to `ISemanticVersion`.
 * Removed all deprecated code.
 * Removed support for mods with no `Name`, `Version`, or `UniqueID` in their manifest.
 * Removed support for mods with a non-unique `UniqueID` value in their manifest.